feat(integrations/github-v2): Add `github-v2` integration scaffold
This patch moves the my scaffold for the new ocean's github integration
into the monorepo structure.

The files and folders for the project have been moved from the my project's
original location to the `integrations/github-v2` directory.
This step is to help align with the library's development practices and thus help me,
begin development of the new integration within the monorepo.